Organizes forces and parsers into separate modules
