+++ open3SPN2/ff3SPN2.py
@@ -10,20 +10,20 @@ It also contains Protein-DNA interaction potentials to be used with openAWSEM.
 __author__ = 'Carlos Bueno'
 __version__ = '0.3.2'
 
-import os
 import configparser
-import subprocess
-import itertools
 import numpy as np
 
 import openmm.app
 import openmm
 import openmm.unit as unit
 import scipy.spatial.distance as sdist
-import pdbfixer
+
 import pandas
-from .geometry import Transform
 from pathlib import Path
+
+from .geometry import Transform
+from .force import *
+from .parser import *
 
 __location__ = Path(__file__).parent.resolve()
 _ef = 1 * unit.kilocalorie / unit.kilojoule  # energy scaling factor
@@ -33,135 +33,6 @@ _complement = {'A': 'T', 'T': 'A', 'G': 'C', 'C': 'G'}
 _dnaResidues = ['DA', 'DC', 'DT', 'DG']
 _proteinResidues = ['IPR', 'IGL', 'NGP']
 xml = __location__/'3SPN2.xml'
-
-
-def parseConfigTable(config_section):
-    """ Parses a section of the configuration file as a table.
-        This function is used to parse the 3SPN2.conf file"""
-
-    def readData(config_section, a):
-        """Filters comments and returns values as a list"""
-        temp = config_section.get(a).split('#')[0].split()
-        l = []
-        for val in temp:
-            val = val.strip()
-            try:
-                x = int(val)
-                l += [x]
-            except ValueError:
-                try:
-                    y = float(val)
-                    l += [y]
-                except ValueError:
-                    l += [val]
-        return l
-
-    data = []
-    for a in config_section:
-        if a == 'name':
-            columns = readData(config_section, a)
-        elif len(a) > 3 and a[:3] == 'row':
-            data += [readData(config_section, a)]
-        else:
-            print(f'Unexpected row {readData(config_section, a)}')
-    return pandas.DataFrame(data, columns=columns)
-
-
-def parsePDB(pdb_file):
-    """ Transforms the pdb file into a pandas table for easy access and data editing"""
-
-    def pdb_line(line):
-        return dict(recname=str(line[0:6]).strip(),
-                    serial=int(line[6:11]),
-                    name=str(line[12:16]).strip(),
-                    altLoc=str(line[16:17]),
-                    resname=str(line[17:20]).strip(),
-                    chainID=str(line[21:22]),
-                    resSeq=int(line[22:26]),
-                    iCode=str(line[26:27]),
-                    x=float(line[30:38]),
-                    y=float(line[38:46]),
-                    z=float(line[46:54]),
-                    occupancy=1.0 if line[54:60].strip()=='' else float(line[54:60]), # Assume occupancy 1 if empty
-                    tempFactor=1.0 if line[60:66].strip()=='' else float(line[60:66]),# Assume beta 1 if empty
-                    element=str(line[76:78]),
-                    charge=str(line[78:80]))
-
-    with open(pdb_file, 'r') as pdb:
-        lines = []
-        for line in pdb:
-            if len(line) > 6 and line[:6] in ['ATOM  ', 'HETATM']:
-                lines += [pdb_line(line)]
-    pdb_atoms = pandas.DataFrame(lines)
-    pdb_atoms = pdb_atoms[['recname', 'serial', 'name', 'altLoc',
-                           'resname', 'chainID', 'resSeq', 'iCode',
-                           'x', 'y', 'z', 'occupancy', 'tempFactor',
-                           'element', 'charge']]
-    return pdb_atoms
-
-
-def fixPDB(pdb_file):
-    """Uses the pdbfixer library to fix a pdb file, replacing non standard residues, removing
-    hetero-atoms and adding missing hydrogens. The input is a pdb file location,
-    the output is a fixer object, which is a pdb in the openawsem format."""
-    fixer = pdbfixer.PDBFixer(filename=pdb_file, )
-    fixer.findMissingResidues()
-    chains = list(fixer.topology.chains())
-    keys = fixer.missingResidues.keys()
-    for key in list(keys):
-        chain_tmp = chains[key[0]]
-        if key[1] in [0, len(list(chain_tmp.residues()))]:
-            del fixer.missingResidues[key]
-
-    fixer.findNonstandardResidues()
-    fixer.replaceNonstandardResidues()
-    fixer.removeHeterogens(keepWater=False)
-    fixer.findMissingAtoms()
-    fixer.addMissingAtoms()
-    fixer.addMissingHydrogens(7.0)
-    return fixer
-
-
-def pdb2table(pdb):
-    """ Parses a pdb in the openmm format and outputs a table that contains all the information
-    on a pdb file """
-    cols = ['recname', 'serial', 'name', 'altLoc',
-            'resname', 'chainID', 'resSeq', 'iCode',
-            'x', 'y', 'z', 'occupancy', 'tempFactor',
-            'element', 'charge']
-    data = []
-    for atom, pos in zip(pdb.topology.atoms(), pdb.positions):
-        residue = atom.residue
-        chain = residue.chain
-        pos = pos.value_in_unit(unit.angstrom)
-        data += [dict(zip(cols, ['ATOM', int(atom.id), atom.name, '',
-                                 residue.name, chain.id, int(residue.id), '',
-                                 pos[0], pos[1], pos[2], 0, 0,
-                                 atom.element.symbol, '']))]
-    atom_list = pandas.DataFrame(data)
-    atom_list = atom_list[cols]
-    atom_list.index = atom_list['serial']
-    return atom_list
-
-def writePDB(atoms, pdb_file='clean.pdb'):
-    """ Writes a minimal version of the pdb file needed for openmm """
-    # Compute chain field
-    if type(atoms['chainID'].iloc[0]) is not str:
-        chain_ix = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz'
-        atoms['chainID'] = [chain_ix[i - 1] for i in atoms['chainID']]
-
-    #Complete missing fields
-    for field, value in {'occupancy':0.0, 'beta':0.0, 'segment':'', 'charge':'', 'insertion':''}.items():
-        if field not in atoms:
-            atoms[field]=value
-
-    # Write pdb file
-    with open(pdb_file, 'w+') as pdb:
-        for i, atom in atoms.iterrows():
-            pdb_line = f'ATOM  {i + 1:>5} {atom["name"]:^4} {atom.resname:<3} {atom.chainID:1}{atom.resSeq:>4}{atom.insertion:1}   {atom.x:>8.3f}{atom.y:>8.3f}{atom.z:>8.3f}{atom.occupancy:>6.2f}{atom.beta:>6.2f}      {atom.segment:4}{atom.element:2}{atom.charge:2}'
-            assert len(pdb_line) == 80, 'An item in the atom table is longer than expected'
-            pdb.write(pdb_line + '\n')
-    return pdb_file
 
 class DNA(object):
     """ A Coarse Grained DNA object."""
@@ -816,932 +687,6 @@ class System(openmm.System):
             energies += [energy]
         return np.array(energies)
 
-
-class Force(object):
-    """ Wrapper for the openMM force. """
-
-    def __init__(self, dna, OpenCLPatch=True):
-        self.periodic = dna.periodic
-        self.force = None
-        self.dna = dna
-        # The patch allows the crosstacking force to run in OpenCL
-        # introducing a small difference in the crosstacking energy
-        self.OpenCLPatch = OpenCLPatch
-
-        # Define the dna force
-        self.reset()
-
-        # Define the interaction pairs
-        self.defineInteraction()
-
-    def __getattr__(self, attr):
-        if attr in self.__dict__:
-            return getattr(self, attr)
-        elif 'force' in self.__dict__:
-            return getattr(self.force, attr)
-        else:
-            if '__repr__' in self.__dict__:
-                raise AttributeError(f"type object {str(self)} has no attribute {str(attr)}")
-            else:
-                raise AttributeError()
-
-    def computeEnergy(self, system, trajectory):
-        # Parse trajectory
-        traj = parse_xyz('Tests/adna/traj.xyz')
-
-        # clear all forces on the system
-        system.clearForces()
-        # setup the force
-        self.setUpInteraction()
-        # for each item of the table:
-        # add the force item
-
-        # compute the energy for every frame
-
-        # return a Table with the energy
-
-    def computeSingleEnergy(self, system, trajectory):
-        # Parse trajectory
-        traj = parse_xyz('Tests/adna/traj.xyz')
-        # for each item of the table:
-        # clear all forces on the system
-        # system.clearForces()
-        # setup the force
-        # self.setUpInteraction()
-        # add the force item
-
-        # compute the energy for every frame
-
-        # return a table with the energy
-
-
-class Bond(Force, openmm.CustomBondForce):
-    def __init__(self, dna, force_group=6, OpenCLPatch=True):
-        self.force_group = force_group
-        super().__init__(dna, OpenCLPatch=OpenCLPatch)
-
-    def getParameterNames(self):
-        self.perInteractionParameters = []
-        self.GlobalParameters = []
-        for i in range(self.force.getNumPerBondParameters()):
-            self.perInteractionParameters += [self.force.getPerBondParameterName(i)]
-        for i in range(self.force.getNumGlobalParameters()):
-            self.GlobalParameters += [self.force.getGlobalParameterName(i)]
-        return [self.perInteractionParameters, self.GlobalParameters]
-
-    def reset(self):
-        bondForce = openmm.CustomBondForce("Kb2*(r-r0)^2+Kb3*(r-r0)^3+Kb4*(r-r0)^4")
-        bondForce.addPerBondParameter('r0')
-        bondForce.addPerBondParameter('Kb2')
-        bondForce.addPerBondParameter('Kb3')
-        bondForce.addPerBondParameter('Kb4')
-        bondForce.setUsesPeriodicBoundaryConditions(self.periodic)
-        bondForce.setForceGroup(self.force_group)
-        self.force = bondForce
-
-    def defineInteraction(self):
-        for i, b in self.dna.bonds.iterrows():
-            # Units converted from
-            parameters = [b['r0'],
-                          b['Kb2'],
-                          b['Kb3'],
-                          b['Kb4']]
-            self.force.addBond(int(b['aai']), int(b['aaj']), parameters)
-
-
-class Angle(Force, openmm.HarmonicAngleForce):
-
-    def __init__(self, dna, force_group=7, OpenCLPatch=True):
-        self.force_group = force_group
-        super().__init__(dna, OpenCLPatch=OpenCLPatch)
-
-    def reset(self):
-        angleForce = openmm.HarmonicAngleForce()
-        angleForce.setUsesPeriodicBoundaryConditions(self.periodic)
-        angleForce.setForceGroup(self.force_group)
-        self.force = angleForce
-
-    def defineInteraction(self):
-        for i, a in self.dna.angles.iterrows():
-            parameters = [a['t0'] * _af,
-                          a['epsilon'] * 2]
-            self.force.addAngle(int(a['aai']), int(a['aaj']), int(a['aak']), *parameters)
-
-
-class Stacking(Force, openmm.CustomCompoundBondForce):
-    def __init__(self, dna, force_group=8, OpenCLPatch=True):
-        self.force_group = force_group
-        super().__init__(dna, OpenCLPatch=OpenCLPatch)
-
-    def reset(self):
-        stackingForce = openmm.CustomCompoundBondForce(3, """energy;
-                        energy=rep+f2*attr;
-                        rep=epsilon*(1-exp(-alpha*(dr)))^2*step(-dr);
-                        attr=epsilon*(1-exp(-alpha*(dr)))^2*step(dr)-epsilon;
-                        dr=distance(p2,p3)-sigma;
-                        f2=max(f*pair2,pair1);
-                        pair1=step(dt+pi/2)*step(pi/2-dt);
-                        pair2=step(dt+pi)*step(pi-dt);
-                        f=1-cos(dt)^2;
-                        dt=rng*(angle(p1,p2,p3)-t0);""")
-        stackingForce.setUsesPeriodicBoundaryConditions(self.periodic)
-        stackingForce.addPerBondParameter('epsilon')
-        stackingForce.addPerBondParameter('sigma')
-        stackingForce.addPerBondParameter('t0')
-        stackingForce.addPerBondParameter('alpha')
-        stackingForce.addPerBondParameter('rng')
-        stackingForce.addGlobalParameter('pi', np.pi)
-        stackingForce.setForceGroup(self.force_group)
-        self.force = stackingForce
-
-    def defineInteraction(self):
-        for i, a in self.dna.stackings.iterrows():
-            parameters = [a['epsilon'],
-                          a['sigma'],
-                          a['t0'] * _af,
-                          a['alpha'],
-                          a['rng']]
-            self.force.addBond([a['aai'], a['aaj'], a['aak']], parameters)
-
-
-class Dihedral(Force, openmm.CustomTorsionForce):
-    def __init__(self, dna, force_group=9, OpenCLPatch=True):
-        self.force_group = force_group
-        super().__init__(dna, OpenCLPatch=OpenCLPatch)
-
-    def reset(self):
-        dihedralForce = openmm.CustomTorsionForce("""energy;
-                        energy = K_periodic*(1-cs)-K_gaussian*exp(-dt_periodic^2/2/sigma^2);
-                        cs = cos(dt);
-                        dt_periodic = dt-floor((dt+pi)/(2*pi))*(2*pi);
-                        dt = theta-t0""")
-        # dihedralForce=simtk.openmm.CustomTorsionForce("theta/60.")
-        dihedralForce.setUsesPeriodicBoundaryConditions(self.periodic)
-        dihedralForce.addPerTorsionParameter('K_periodic')
-        dihedralForce.addPerTorsionParameter('K_gaussian')
-        dihedralForce.addPerTorsionParameter('sigma')
-        dihedralForce.addPerTorsionParameter('t0')
-        dihedralForce.addGlobalParameter('pi', np.pi)
-        dihedralForce.setForceGroup(self.force_group)
-        self.force = dihedralForce
-
-    def defineInteraction(self):
-        for i, a in self.dna.dihedrals.iterrows():
-            parameters = [a['K_dihedral'],
-                          a['K_gaussian'],
-                          a['sigma'],
-                          (180 + a['t0']) * _af]
-            particles = [a['aai'], a['aaj'], a['aak'], a['aal']]
-            self.force.addTorsion(*particles, parameters)
-
-
-class BasePair(Force, openmm.CustomHbondForce):
-    def __init__(self, dna, force_group=10, OpenCLPatch=True):
-        self.force_group = force_group
-        super().__init__(dna, OpenCLPatch=OpenCLPatch)
-
-    def reset(self):
-        def basePairForce():
-            pairForce = openmm.CustomHbondForce('''energy;
-                        energy=rep+1/2*(1+cos(dphi))*fdt1*fdt2*attr;
-                        rep  = epsilon*(1-exp(-alpha*dr))^2*(1-step(dr));
-                        attr = epsilon*(1-exp(-alpha*dr))^2*step(dr)-epsilon;
-                        fdt1 = max(f1*pair0t1,pair1t1);
-                        fdt2 = max(f2*pair0t2,pair1t2);
-                        pair1t1 = step(pi/2+dt1)*step(pi/2-dt1);
-                        pair1t2 = step(pi/2+dt2)*step(pi/2-dt2);
-                        pair0t1 = step(pi+dt1)*step(pi-dt1);
-                        pair0t2 = step(pi+dt2)*step(pi-dt2);
-                        f1 = 1-cos(dt1)^2;
-                        f2 = 1-cos(dt2)^2;
-                        dphi = dihedral(d2,d1,a1,a2)-phi0;
-                        dr    = distance(d1,a1)-sigma;
-                        dt1   = rng*(angle(d2,d1,a1)-t01);
-                        dt2   = rng*(angle(a2,a1,d1)-t02);''')
-            if self.periodic:
-                pairForce.setNonbondedMethod(pairForce.CutoffPeriodic)
-            else:
-                pairForce.setNonbondedMethod(pairForce.CutoffNonPeriodic)
-            pairForce.setCutoffDistance(1.8)  # Paper
-            pairForce.addPerDonorParameter('phi0')
-            pairForce.addPerDonorParameter('sigma')
-            pairForce.addPerDonorParameter('t01')
-            pairForce.addPerDonorParameter('t02')
-            pairForce.addPerDonorParameter('rng')
-            pairForce.addPerDonorParameter('epsilon')
-            pairForce.addPerDonorParameter('alpha')
-            pairForce.addGlobalParameter('pi', np.pi)
-            self.force = pairForce
-            pairForce.setForceGroup(self.force_group)
-            return pairForce
-
-        basePairForces = {}
-        pair_definition = self.dna.pair_definition[self.dna.pair_definition['DNA'] == self.dna.DNAtype]
-        for i, pair in pair_definition.iterrows():
-            basePairForces.update({i: basePairForce()})
-        self.forces = basePairForces
-
-    def defineInteraction(self):
-        pair_definition = self.dna.pair_definition[self.dna.pair_definition['DNA'] == self.dna.DNAtype]
-        atoms = self.dna.atoms.copy()
-        atoms['index'] = atoms.index
-        atoms.index = zip(atoms['chainID'], atoms['resSeq'], atoms['name'])
-        is_dna = atoms['resname'].isin(_dnaResidues)
-
-        for i, pair in pair_definition.iterrows():
-            D1 = atoms[(atoms['name'] == pair['Base1']) & is_dna].copy()
-            A1 = atoms[(atoms['name'] == pair['Base2']) & is_dna].copy()
-
-            try:
-                D2 = atoms.loc[[(c, r, 'S') for c, r, n in D1.index]]
-            except KeyError:
-                for c, r, n in D1.index:
-                    if (c, r, 'S') not in atoms.index:
-                        print(f'Residue {c}:{r} does not have a Sugar atom (S)')
-                raise KeyError
-
-            try:
-                A2 = atoms.loc[[(c, r, 'S') for c, r, n in A1.index]]
-            except KeyError:
-                for c, r, n in A1.index:
-                    if (c, r, 'S') not in atoms.index:
-                        print(f'Residue {c}:{r} does not have a Sugar atom (S)')
-                raise KeyError
-
-            D1_list = list(D1['index'])
-            A1_list = list(A1['index'])
-            D2_list = list(D2['index'])
-            A2_list = list(A2['index'])
-
-            # Define parameters
-            parameters = [pair.torsion * _af,
-                          pair.sigma,
-                          pair.t1 * _af,
-                          pair.t2 * _af,
-                          pair.rang,
-                          pair.epsilon,
-                          pair.alpha]
-
-            # Add donors and acceptors
-            # Here I am including the same atom twice,
-            # it doesn't seem to break things
-            for d1, d2 in zip(D1_list, D2_list):
-                self.forces[i].addDonor(d1, d2, -1, parameters)
-                #print(d1, d2, d2, parameters)
-            for a1, a2 in zip(A1_list, A2_list):
-                self.forces[i].addAcceptor(a1, a2, -1)
-                #print(a1, a2, a2)
-            # Exclude interactions
-            D1['donor_id'] = [i for i in range(len(D1))]
-            A1['aceptor_id'] = [i for i in range(len(A1))]
-
-            for (_i, atom_a), (_j, atom_b) in itertools.product(D1.iterrows(), A1.iterrows()):
-                # Neighboring residues
-                # The sequence exclusion was reduced to two residues
-                # since the maximum number of exclusions in OpenCL is 4.
-                # In the original 3SPN2 it was 3 residues (6 to 9)
-                # This change has no noticeable effect
-                if (atom_a.chainID == atom_b.chainID) and (abs(atom_a.resSeq - atom_b.resSeq) <= 2):
-                    self.forces[i].addExclusion(atom_a['donor_id'], atom_b['aceptor_id'])
-                    #print(_i, _j)
-
-    def addForce(self, system):
-        for f in self.forces:
-            system.addForce(self.forces[f])
-
-
-class CrossStacking(Force):
-    def __init__(self, dna, force_group=11, OpenCLPatch=True):
-        self.force_group = force_group
-        super().__init__(dna, OpenCLPatch=OpenCLPatch)
-
-    def reset(self):
-        def crossStackingForce(parametersOnDonor=False):
-            crossForce = openmm.CustomHbondForce(f'''energy;
-                         energy   = fdt3*fdtCS*attr/2;
-                         attr     = epsilon*(1-exp(-alpha*dr))^2*step(dr)-epsilon;
-                         fdt3     = max(f1*pair0t3,pair1t3);
-                         fdtCS    = max(f2*pair0tCS,pair1tCS);
-                         pair0t3  = step(pi+dt3)*step(pi-dt3);
-                         pair0tCS = step(pi+dtCS)*step(pi-dtCS);
-                         pair1t3  = step(pi/2+dt3)*step(pi/2-dt3);
-                         pair1tCS = step(pi/2+dtCS)*step(pi/2-dtCS);
-                         f1       = 1-cos(dt3)^2;
-                         f2       = 1-cos(dtCS)^2;
-                         dr       = distance(d1,a3)-sigma;
-                         dt3      = rng_BP*(t3-t03);
-                         dtCS     = rng_CS*(tCS-t0CS);
-                         tCS      = angle(d2,d1,a3);
-                         t3       = acos(cost3lim);
-                         cost3lim = min(max(cost3,-0.99),0.99);
-                         cost3    = sin(t1)*sin(t2)*cos(phi)-cos(t1)*cos(t2);
-                         t1       = angle(d2,d1,a1);
-                         t2       = angle(d1,a1,a2);
-                         phi      = dihedral(d2,d1,a1,a2);''')
-            if self.periodic:
-                crossForce.setNonbondedMethod(crossForce.CutoffPeriodic)
-            else:
-                crossForce.setNonbondedMethod(crossForce.CutoffNonPeriodic)
-            crossForce.setCutoffDistance(1.8)  # Paper
-            parameters = ['t03', 't0CS', 'rng_CS', 'rng_BP', 'epsilon', 'alpha', 'sigma']
-            for p in parameters:
-                if parametersOnDonor:
-                    crossForce.addPerDonorParameter(p)
-                else:
-                    crossForce.addPerAcceptorParameter(p)
-            crossForce.addGlobalParameter('pi', np.pi)
-            crossForce.setForceGroup(self.force_group)
-            return crossForce
-
-        crossStackingForces = {}
-        for base in ['A', 'T', 'G', 'C']:
-            crossStackingForces.update({base: (crossStackingForce(), crossStackingForce())})
-        self.crossStackingForces = crossStackingForces
-
-    def defineInteraction(self):
-        atoms = self.dna.atoms.copy()
-        atoms['index'] = atoms.index
-        atoms.index = zip(atoms['chainID'], atoms['resSeq'], atoms['name'].replace(['A', 'C', 'T', 'G'], 'B'))
-        is_dna = atoms['resname'].isin(_dnaResidues)
-        bases = atoms[atoms['name'].isin(['A', 'T', 'G', 'C']) & is_dna]
-        D1 = bases
-        D2 = atoms.reindex([(c, r, 'S') for c, r, n in bases.index])
-        D3 = atoms.reindex([(c, r + 1, 'B') for c, r, n in bases.index])
-        A1 = D1
-        A2 = D2
-        A3 = atoms.reindex([(c, r - 1, 'B') for c, r, n in bases.index])
-
-        # Select only bases where the other atoms exist
-        D2.index = D1.index
-        D3.index = D1.index
-        temp = pandas.concat([D1, D2, D3], axis=1, keys=['D1', 'D2', 'D3'])
-        sel = temp[temp['D3', 'name'].isin(['A', 'T', 'G', 'C']) &  # D3 must be a base
-                   temp['D2', 'name'].isin(['S']) &  # D2 must be a sugar
-                   (temp['D3', 'chainID'] == temp['D1', 'chainID']) &  # D3 must be in the same chain
-                   (temp['D2', 'chainID'] == temp['D1', 'chainID'])].index  # D2 must be in the same chain
-        D1 = atoms.reindex(sel)
-        D2 = atoms.reindex([(c, r, 'S') for c, r, n in sel])
-        D3 = atoms.reindex([(c, r + 1, 'B') for c, r, n in sel])
-
-        # Aceptors
-        A2.index = A1.index
-        A3.index = A1.index
-        temp = pandas.concat([A1, A2, A3], axis=1, keys=['A1', 'A2', 'A3'])
-        sel = temp[temp['A3', 'name'].isin(['A', 'T', 'G', 'C']) &  # A3 must be a base
-                   temp['A2', 'name'].isin(['S']) &  # A2 must be a sugar
-                   (temp['A3', 'chainID'] == temp['A1', 'chainID']) &  # A3 must be in the same chain
-                   (temp['A2', 'chainID'] == temp['A1', 'chainID'])].index  # A2 must be in the same chain
-        A1 = atoms.reindex(sel)
-        A2 = atoms.reindex([(c, r, 'S') for c, r, n in sel])
-        A3 = atoms.reindex([(c, r - 1, 'B') for c, r, n in sel])
-
-        # Parameters
-        cross_definition = self.dna.cross_definition[self.dna.cross_definition['DNA'] == self.dna.DNAtype].copy()
-        i = [a for a in zip(cross_definition['Base_d1'], cross_definition['Base_a1'], cross_definition['Base_a3'])]
-        cross_definition.index = i
-
-        donors = {i: [] for i in ['A', 'T', 'G', 'C']}
-        for donator, donator2, d1, d2, d3 in zip(D1.itertuples(), D3.itertuples(), D1['index'], D2['index'],
-                                                 D3['index']):
-            d1t = donator.name
-            d3t = donator2.name
-            c1, c2 = self.crossStackingForces[d1t]
-            a1t = _complement[d1t]
-            param = cross_definition.loc[[(a1t, d1t, d3t)]].squeeze()
-            parameters = [param['t03'] * _af,
-                          param['T0CS_2'] * _af,
-                          param['rng_cs2'],
-                          param['rng_bp'],
-                          param['eps_cs2'],
-                          param['alpha_cs2'],
-                          param['Sigma_2']]
-            c1.addDonor(d1, d2, d3)
-            c2.addAcceptor(d1, d2, d3, parameters)
-            # print("Donor", d1t, d1, d2, d3)
-            donors[d1t] += [d1]
-
-        aceptors = {i: [] for i in ['A', 'T', 'G', 'C']}
-        for aceptor, aceptor2, a1, a2, a3 in zip(A1.itertuples(), A3.itertuples(), A1['index'], A2['index'],
-                                                 A3['index']):
-            a1t = aceptor.name
-            a3t = aceptor2.name
-            c1, c2 = self.crossStackingForces[_complement[a1t]]
-            d1t = _complement[a1t]
-            param = cross_definition.loc[[(d1t, a1t, a3t)]].squeeze()
-            parameters = [param['t03'] * _af,
-                          param['T0CS_1'] * _af,
-                          param['rng_cs1'],
-                          param['rng_bp'],
-                          param['eps_cs1'],
-                          param['alpha_cs1'],
-                          param['Sigma_1']]
-            c1.addAcceptor(a1, a2, a3, parameters)
-            c2.addDonor(a1, a2, a3)
-            # print("Aceptor", a1t, a1, a2, a3)
-            aceptors[_complement[a1t]] += [a1]
-
-        # Exclusions
-        for base in ['A', 'T', 'G', 'C']:
-            c1, c2 = self.crossStackingForces[base]
-            for ii, i in enumerate(donors[base]):
-                for jj, j in enumerate(aceptors[base]):
-                    # The sequence exclusion was reduced to two residues
-                    # since the maximum number of exclusions in OpenCL is 4.
-                    # In the original 3SPN2 it was 3 residues (6 to 9)
-                    # This change has a small effect in B-DNA and curved B-DNA
-                    # The second change is to make the interaction symetric and dividing the energy over 2
-                    # This also reduces the number of exclusions in the force
-                    maxn = 6 if self.OpenCLPatch else 9
-                    if (self.dna.atoms.at[i, 'chainID'] == self.dna.atoms.at[j, 'chainID'] and abs(i - j) <= maxn) or \
-                            (not self.OpenCLPatch and i > j):
-                        c1.addExclusion(ii, jj)
-                        c2.addExclusion(jj, ii)
-
-    def addForce(self, system):
-        for c1, c2 in self.crossStackingForces.values():
-            system.addForce(c1)
-            system.addForce(c2)
-
-    def getForceGroup(self):
-        fg = 0
-        for c1, c2 in self.crossStackingForces.values():
-            fg = c1.getForceGroup()
-            break
-        for c1, c2 in self.crossStackingForces.values():
-            assert fg == c1.getForceGroup()
-            assert fg == c2.getForceGroup()
-        return fg
-
-def addNonBondedExclusions(dna, force, OpenCLPatch=True):
-    is_dna = dna.atoms['resname'].isin(_dnaResidues)
-    atoms = dna.atoms.copy()
-    selection = atoms[is_dna].sort_index()
-    selection['index'] = selection.index
-    selection['neighbor'] = selection['chainID'].astype(str) + '_' + (selection['resSeq'] - 1).astype(str)
-    selection.index = selection['chainID'].astype(str) + '_' + (selection['resSeq']).astype(str)
-
-    exclusions = []
-    for i, neighbor_res, self_res in zip(selection['index'], selection['neighbor'], selection.index):
-        # Add exclusions for the same residue
-        for j in selection.loc[self_res, 'index']:
-            if i > j:
-                exclusions += [(j, i)]
-        # Add exclusions with the neighboring residue on the same chain
-        try:
-            for j in selection.loc[neighbor_res, 'index']:
-                exclusions += [(j, i)]
-        except KeyError:
-            continue
-
-    exclusions_bp = []
-    if OpenCLPatch:
-        # Add basepair exclusions
-        for key in _complement:
-            selection_N = atoms[is_dna & (atoms['name'] == key)]
-            selection_C = atoms[is_dna & (atoms['name'] == _complement[key])]
-            for i, j in itertools.product(selection_N.index, selection_C.index):
-                if i > j:
-                    exclusions_bp += [(j, i)]
-
-    exclusions = list(set(exclusions+exclusions_bp))
-    for i, j in exclusions:
-        force.addExclusion(i, j)
-
-
-class Exclusion(Force, openmm.CustomNonbondedForce):
-    def __init__(self, dna, force_group = 12, OpenCLPatch=True):
-        self.force_group = force_group
-        super().__init__(dna, OpenCLPatch=OpenCLPatch)
-
-    def reset(self):
-        exclusionForce = openmm.CustomNonbondedForce("""energy;
-                         energy=(epsilon*((sigma/r)^12-2*(sigma/r)^6)+epsilon)*step(sigma-r);
-                         sigma=0.5*(sigma1+sigma2);
-                         epsilon=sqrt(epsilon1*epsilon2)""")
-        exclusionForce.addPerParticleParameter('epsilon')
-        exclusionForce.addPerParticleParameter('sigma')
-        exclusionForce.setCutoffDistance(1.8)
-        exclusionForce.setForceGroup(self.force_group)  # There can not be multiple cutoff distance on the same force group
-        if self.periodic:
-            exclusionForce.setNonbondedMethod(exclusionForce.CutoffPeriodic)
-        else:
-            exclusionForce.setNonbondedMethod(exclusionForce.CutoffNonPeriodic)
-        self.force = exclusionForce
-
-    def defineInteraction(self):
-        # addParticles
-        particle_definition = self.dna.particle_definition[self.dna.particle_definition['DNA'] == self.dna.DNAtype]
-        particle_definition.index = particle_definition.name
-
-        # Reduces or increases the cutoff to the maximum particle radius
-        self.force.setCutoffDistance(particle_definition.radius.max())
-
-        # Select only dna atoms
-        is_dna = self.dna.atoms['resname'].isin(_dnaResidues)
-        atoms = self.dna.atoms.copy()
-        atoms['is_dna'] = is_dna
-        for i, atom in atoms.iterrows():
-            if atom.is_dna:
-                param = particle_definition.loc[atom['name']]
-                parameters = [param.epsilon,
-                              param.radius]
-            else:
-                parameters = [0, .1]  # Null energy and some radius)
-            # print(i, parameters)
-            self.force.addParticle(parameters)
-
-        # addExclusions
-        addNonBondedExclusions(self.dna, self.force)
-
-
-class Electrostatics(Force, openmm.CustomNonbondedForce):
-    def __init__(self, dna, force_group=13, temperature=300*unit.kelvin, salt_concentration=100*unit.millimolar, OpenCLPatch=True):
-        self.force_group = force_group
-        self.T = temperature
-        self.C = salt_concentration
-        super().__init__(dna, OpenCLPatch=OpenCLPatch)
-
-    def reset(self):
-        T = self.T
-        C = self.C
-        e = 249.4 - 0.788 * (T / unit.kelvin) + 7.2E-4 * (T / unit.kelvin) ** 2
-        a = 1 - 0.2551 * (C / unit.molar) + 5.151E-2 * (C / unit.molar) ** 2 - 6.889E-3 * (C / unit.molar) ** 3
-        #print(e, a)
-        dielectric = e * a
-        # Debye length
-        kb = unit.BOLTZMANN_CONSTANT_kB  # Bolztmann constant
-        Na = unit.AVOGADRO_CONSTANT_NA  # Avogadro number
-        ec = 1.60217653E-19 * unit.coulomb  # proton charge
-        pv = 8.8541878176E-12 * unit.farad / unit.meter  # dielectric permittivity of vacuum
-
-        ldby = np.sqrt(dielectric * pv * kb * T / (2.0 * Na * ec ** 2 * C))
-        ldby = ldby.in_units_of(unit.nanometer)
-        denominator = 4 * np.pi * pv * dielectric / (Na * ec ** 2)
-        denominator = denominator.in_units_of(unit.kilocalorie_per_mole**-1 * unit.nanometer**-1)
-        #print(ldby, denominator)
-
-        electrostaticForce = openmm.CustomNonbondedForce("""energy;
-                                                                energy=q1*q2*exp(-r/dh_length)/denominator/r;""")
-        electrostaticForce.addPerParticleParameter('q')
-        electrostaticForce.addGlobalParameter('dh_length', ldby)
-        electrostaticForce.addGlobalParameter('denominator', denominator)
-
-        electrostaticForce.setCutoffDistance(5)
-        if self.periodic:
-            electrostaticForce.setNonbondedMethod(electrostaticForce.CutoffPeriodic)
-        else:
-            electrostaticForce.setNonbondedMethod(electrostaticForce.CutoffNonPeriodic)
-        electrostaticForce.setForceGroup(self.force_group)
-        self.force = electrostaticForce
-
-    def defineInteraction(self):
-        # addParticles
-        particle_definition = self.dna.particle_definition[self.dna.particle_definition['DNA'] == self.dna.DNAtype]
-        particle_definition.index = particle_definition.name
-
-        # Select only dna atoms
-        is_dna = self.dna.atoms['resname'].isin(_dnaResidues)
-        atoms = self.dna.atoms.copy()
-        atoms['is_dna'] = is_dna
-
-        for i, atom in atoms.iterrows():
-            if atom.is_dna:
-                param = particle_definition.loc[atom['name']]
-                parameters = [param.charge]
-            else:
-                parameters = [0]  # No charge if it is not DNA
-            # print (i,parameters)
-            self.force.addParticle(parameters)
-
-        # add neighbor exclusion
-        addNonBondedExclusions(self.dna, self.force, self.OpenCLPatch)
-
-
-class ProteinDNAForce(Force):
-    def __init__(self, dna, protein):
-        self.protein = protein
-        super().__init__(dna)
-
-
-class ExclusionProteinDNA(ProteinDNAForce):
-    """ Protein-DNA exclusion potential"""
-    def __init__(self, dna, protein, k=1, force_group=14):
-        self.k = k
-        self.force_group = force_group
-        super().__init__(dna, protein)
-
-    def reset(self):
-        exclusionForce = openmm.CustomNonbondedForce(f"""k_exclusion_protein_DNA*energy;
-                         energy=(4*epsilon*((sigma/r)^12-(sigma/r)^6)-offset)*step(cutoff-r);
-                         offset=4*epsilon*((sigma/cutoff)^12-(sigma/cutoff)^6);
-                         sigma=0.5*(sigma1+sigma2); 
-                         epsilon=sqrt(epsilon1*epsilon2);
-                         cutoff=sqrt(cutoff1*cutoff2)""")
-        exclusionForce.addGlobalParameter('k_exclusion_protein_DNA', self.k)
-        exclusionForce.addPerParticleParameter('epsilon')
-        exclusionForce.addPerParticleParameter('sigma')
-        exclusionForce.addPerParticleParameter('cutoff')
-        exclusionForce.setCutoffDistance(1.55)
-        # exclusionForce.setUseLongRangeCorrection(True)
-        exclusionForce.setForceGroup(self.force_group)  # There can not be multiple cutoff distance on the same force group
-        if self.periodic:
-            exclusionForce.setNonbondedMethod(exclusionForce.CutoffPeriodic)
-        else:
-            exclusionForce.setNonbondedMethod(exclusionForce.CutoffNonPeriodic)
-        self.force = exclusionForce
-
-    def defineInteraction(self):
-
-        particle_definition = self.dna.config['Protein-DNA particles']
-        dna_particle_definition=particle_definition[(particle_definition['molecule'] == 'DNA') &
-                                                    (particle_definition['DNA'] == self.dna.DNAtype)]
-        protein_particle_definition = particle_definition[(particle_definition['molecule'] == 'Protein')]
-
-        # Merge DNA and protein particle definitions
-        particle_definition = pandas.concat([dna_particle_definition, protein_particle_definition], sort=False)
-        particle_definition.index = particle_definition.molecule + particle_definition.name
-        self.particle_definition = particle_definition
-
-        is_dna = self.dna.atoms['resname'].isin(_dnaResidues)
-        is_protein = self.dna.atoms['resname'].isin(_proteinResidues)
-        atoms = self.dna.atoms.copy()
-        atoms['is_dna'] = is_dna
-        atoms['is_protein'] = is_protein
-        atoms['epsilon']=np.nan
-        atoms['radius']=np.nan
-        atoms['cutoff'] = np.nan
-        DNA_list = []
-        protein_list = []
-        for i, atom in atoms.iterrows():
-            if atom.is_dna:
-                param = particle_definition.loc['DNA' + atom['name']]
-                parameters = [param.epsilon,
-                              param.radius,
-                              param.cutoff]
-                DNA_list += [i]
-            elif atom.is_protein:
-                param = particle_definition.loc['Protein' + atom['name']]
-                parameters = [param.epsilon,
-                              param.radius,
-                              param.cutoff]
-                protein_list += [i]
-            else:
-                print(f'Residue {i} not included in protein-DNA interactions')
-                parameters = [0, .1,.1]
-            atoms.loc[i, ['epsilon', 'radius', 'cutoff']] = parameters
-            self.atoms = atoms
-            self.force.addParticle(parameters)
-        self.force.addInteractionGroup(DNA_list, protein_list)
-
-        # addExclusions
-        addNonBondedExclusions(self.dna, self.force)
-
-
-class ElectrostaticsProteinDNA(ProteinDNAForce):
-    """DNA-protein and protein-protein electrostatics."""
-    def __init__(self, dna, protein, k=1, force_group=15):
-        self.k = k
-        self.force_group = force_group
-        super().__init__(dna, protein)
-
-    def reset(self):
-        dielectric = 78 # e * a
-        #print(dielectric)
-        # Debye length
-        Na = unit.AVOGADRO_CONSTANT_NA  # Avogadro number
-        ec = 1.60217653E-19 * unit.coulomb  # proton charge
-        pv = 8.8541878176E-12 * unit.farad / unit.meter  # dielectric permittivity of vacuum
-
-        ldby = 1.2 * unit.nanometer # np.sqrt(dielectric * pv * kb * T / (2.0 * Na * ec ** 2 * C))
-        denominator = 4 * np.pi * pv * dielectric / (Na * ec ** 2)
-        denominator = denominator.in_units_of(unit.kilocalorie_per_mole**-1 * unit.nanometer**-1)
-        #print(ldby, denominator)
-        k = self.k
-        electrostaticForce = openmm.CustomNonbondedForce(f"""k_electro_protein_DNA*energy;
-                             energy=q1*q2*exp(-r/inter_dh_length)/inter_denominator/r;""")
-        electrostaticForce.addPerParticleParameter('q')
-        electrostaticForce.addGlobalParameter('k_electro_protein_DNA', k)
-        electrostaticForce.addGlobalParameter('inter_dh_length', ldby)
-        electrostaticForce.addGlobalParameter('inter_denominator', denominator)
-
-        electrostaticForce.setCutoffDistance(4)
-        if self.periodic:
-            electrostaticForce.setNonbondedMethod(electrostaticForce.CutoffPeriodic)
-        else:
-            electrostaticForce.setNonbondedMethod(electrostaticForce.CutoffNonPeriodic)
-        electrostaticForce.setForceGroup(self.force_group)
-        self.force = electrostaticForce
-
-    def defineInteraction(self):
-        # Merge DNA and protein particle definitions
-        particle_definition = self.dna.config['Protein-DNA particles']
-        dna_particle_definition=particle_definition[(particle_definition['molecule'] == 'DNA') &
-                                                    (particle_definition['DNA'] == self.dna.DNAtype)]
-        protein_particle_definition = particle_definition[(particle_definition['molecule'] == 'Protein')]
-
-        # Merge DNA and protein particle definitions
-        particle_definition = pandas.concat([dna_particle_definition, protein_particle_definition], sort=False)
-        particle_definition.index = particle_definition.molecule + particle_definition.name
-        self.particle_definition = particle_definition
-
-        # Open Sequence dependent electrostatics
-        sequence_electrostatics = self.dna.config['Sequence dependent electrostatics']
-        sequence_electrostatics.index = sequence_electrostatics.resname
-
-        # Select only dna and protein atoms
-        is_dna = self.protein.atoms['resname'].isin(_dnaResidues)
-        is_protein = self.protein.atoms['resname'].isin(_proteinResidues)
-        atoms = self.protein.atoms.copy()
-        atoms['is_dna'] = is_dna
-        atoms['is_protein'] = is_protein
-        DNA_list = []
-        protein_list = []
-
-        for i, atom in atoms.iterrows():
-            if atom.is_dna:
-                param = particle_definition.loc['DNA' + atom['name']]
-                charge = param.charge
-                parameters = [charge]
-                if charge != 0:
-                    DNA_list += [i]
-                    #print(atom.chainID, atom.resSeq, atom.resname, atom['name'], charge)
-            elif atom.is_protein:
-                atom_param = particle_definition.loc['Protein' + atom['name']]
-                seq_param = sequence_electrostatics.loc[atom.real_resname]
-                charge = atom_param.charge * seq_param.charge
-                parameters = [charge]
-                if charge != 0:
-                    protein_list += [i]
-                    #print(atom.chainID, atom.resSeq, atom.resname, atom['name'], charge)
-            else:
-                print(f'Residue {i} not included in protein-DNA electrostatics')
-                parameters = [0]  # No charge if it is not DNA
-            # print (i,parameters)
-            self.force.addParticle(parameters)
-        self.force.addInteractionGroup(DNA_list, protein_list)
-        # self.force.addInteractionGroup(protein_list, protein_list) #protein-protein electrostatics should be included using debye Huckel Terms
-
-        # addExclusions
-        addNonBondedExclusions(self.dna, self.force)
-
-
-class AMHgoProteinDNA(ProteinDNAForce):
-    """ Protein-DNA amhgo potential"""
-    def __init__(self, dna, protein, chain_protein='A', chain_DNA='B', k_amhgo_PD=1*unit.kilocalorie_per_mole, sigma_sq=0.05*unit.nanometers**2, aaweight=False, globalct=True, cutoff=1.8, force_group=16):
-        self.force_group = force_group
-        self.k_amhgo_PD = k_amhgo_PD
-        self.sigma_sq= sigma_sq
-        self.chain_protein = chain_protein
-        self.chain_DNA = chain_DNA
-        self.aaweight = aaweight
-        self.cutoff = cutoff
-        self.globalct = globalct
-        super().__init__(dna, protein)
-
-    def reset(self):
-        cutoff = self.cutoff
-        k_3spn2 = self.k_3spn2
-        if self.globalct:
-                amhgoForce = openmm.CustomBondForce(f"-k_amhgo_PD*gamma_ij*exp(-(r-r_ijN)^2/(2*sigma_sq))*step({cutoff}-r)")
-        else:
-                amhgoForce = openmm.CustomBondForce(f"-k_amhgo_PD*gamma_ij*exp(-(r-r_ijN)^2/(2*sigma_sq))*step(r_ijN+{cutoff}-r)")
-        amhgoForce.addGlobalParameter("k_amhgo_PD", k_3spn2*self.k_amhgo_PD)
-        amhgoForce.addGlobalParameter("sigma_sq", self.sigma_sq)
-        amhgoForce.addPerBondParameter("gamma_ij")
-        amhgoForce.addPerBondParameter("r_ijN")
-        amhgoForce.setUsesPeriodicBoundaryConditions(self.periodic)
-        amhgoForce.setForceGroup(self.force_group)  # There can not be multiple cutoff distance on the same force group
-        self.force = amhgoForce
-
-    def defineInteraction(self):
-        atoms = self.dna.atoms.copy()
-        atoms['index'] = atoms.index
-        atoms.index = zip(atoms['chainID'], atoms['resSeq'], atoms['name'])
-
-        contact_list = np.loadtxt("contact_protein_DNA.dat")
-        for i in range(len(contact_list)):
-            if self.aaweight:
-                gamma_ij = contact_list[i][3]
-            else:
-                gamma_ij = 1.0
-            if (self.chain_protein, int(contact_list[i][0]), 'CB') in atoms.index:
-                 CB_protein = atoms[(atoms['chainID'] == self.chain_protein) & (atoms['resSeq'] == int(contact_list[i][0])) & (atoms['name'] == 'CB') & atoms['resname'].isin(_proteinResidues)].copy()
-            else:
-                 CB_protein = atoms[(atoms['chainID'] == self.chain_protein) & (atoms['resSeq'] == int(contact_list[i][0])) & (atoms['name'] == 'CA') & atoms['resname'].isin(_proteinResidues)].copy()
-            base_DNA = atoms[(atoms['chainID'] == self.chain_DNA) & (atoms['resSeq'] == int(contact_list[i][1])) & (atoms['name'].isin(['A', 'T', 'G', 'C'])) & atoms['resname'].isin(_dnaResidues)].copy()
-            r_ijN = contact_list[i][2]/10.0*unit.nanometers
-            self.force.addBond(int(CB_protein['index'].values[0]), int(base_DNA['index'].values[0]), [gamma_ij, r_ijN])
-            print(int(CB_protein['index'].values[0]), int(base_DNA['index'].values[0]), [gamma_ij, r_ijN])
-
-
-#class AMHgoProteinDNA(ProteinDNAForce):
-#    """ Protein-DNA amhgo potential (Xinyu)"""
-#    def __init__(self, dna, protein, chain_protein='A', chain_DNA='B', k_amhgo_PD=1*unit.kilocalorie_per_mole,
-#                 sigma_sq=0.05*unit.nanometers**2, aaweight=False, cutoff=1.8, force_group=16):
-#        self.force_group = force_group
-#        self.k_amhgo_PD = k_amhgo_PD
-#        self.sigma_sq= sigma_sq
-#        self.chain_protein = chain_protein
-#        self.chain_DNA = chain_DNA
-#        self.aaweight = aaweight
-#        self.cutoff = cutoff
-#        super().__init__(dna, protein)
-#
-#    def reset(self):
-#        cutoff = self.cutoff
-#        amhgoForce = openmm.CustomBondForce(f"-k_amhgo_PD*gamma_ij*exp(-(r-r_ijN)^2/(2*sigma_sq))*step({cutoff}-r)")
-#        amhgoForce.addGlobalParameter("k_amhgo_PD", self.k_amhgo_PD)
-#        amhgoForce.addGlobalParameter("sigma_sq", self.sigma_sq)
-#        amhgoForce.addPerBondParameter("gamma_ij")
-#        amhgoForce.addPerBondParameter("r_ijN")
-#        amhgoForce.setUsesPeriodicBoundaryConditions(self.periodic)
-#        amhgoForce.setForceGroup(self.force_group)  # There can not be multiple cutoff distance on the same force group
-#        self.force = amhgoForce
-#
-#    def defineInteraction(self):
-#        atoms = self.dna.atoms.copy()
-#        atoms['index'] = atoms.index
-#        atoms.index = zip(atoms['chainID'], atoms['resSeq'], atoms['name'])
-#
-#        contact_list = np.loadtxt("contact_protein_DNA.dat")
-#        for i in range(len(contact_list)):
-#            if self.aaweight: gamma_ij = contact_list[i][3]
-#            else:   gamma_ij = 1.0
-#            if (self.chain_protein, int(contact_list[i][0]), 'CB') in atoms.index:
-#                 CB_protein = atoms[(atoms['chainID'] == self.chain_protein) & (atoms['resSeq'] == int(contact_list[i][0])) &
-#                                    (atoms['name'] == 'CB') & atoms['resname'].isin(_proteinResidues)].copy()
-#            else:
-#                 CB_protein = atoms[(atoms['chainID'] == self.chain_protein) & (atoms['resSeq'] == int(contact_list[i][0])) &
-#                                    (atoms['name'] == 'CA') & atoms['resname'].isin(_proteinResidues)].copy()
-#            base_DNA = atoms[(atoms['chainID'] == self.chain_DNA) & (atoms['resSeq'] == int(contact_list[i][1])) & (atoms['name'].isin(['A', 'T', 'G', 'C'])) & atoms['resname'].isin(_dnaResidues)].copy()
-#            r_ijN = contact_list[i][2]/10.0*unit.nanometers
-#            self.force.addBond(int(CB_protein['index'].values[0]), int(base_DNA['index'].values[0]), [gamma_ij, r_ijN])
-#            print(int(CB_protein['index'].values[0]), int(base_DNA['index'].values[0]), [gamma_ij, r_ijN])
-
-class StringProteinDNA(ProteinDNAForce):
-    """ Protein-DNA string potential (Xinyu)"""
-    def __init__(self, dna, protein, r0, chain_protein='A', chain_DNA='B', k_string_PD=10*4.184, protein_seg=False, group=[]):
-        self.k_string_PD = k_string_PD
-        self.chain_protein = chain_protein
-        self.chain_DNA = chain_DNA
-        self.r0 = r0
-        self.protein_seg = protein_seg
-        self.group = group
-        super().__init__(dna, protein)
-
-    def reset(self):
-        r0=self.r0
-        k_string_PD=self.k_string_PD
-        stringForce = openmm.CustomCentroidBondForce(2, f"0.5*{k_string_PD}*(distance(g1,g2)-{r0})^2")
-        self.force = stringForce
-        print("String_PD bias on: r0, k_string = ", r0, k_string_PD)
-
-    def defineInteraction(self):
-        atoms = self.dna.atoms.copy()
-        atoms['index'] = atoms.index
-        CA_atoms = atoms[(atoms['chainID'] == self.chain_protein) & (atoms['name'] == 'CA') & atoms['resname'].isin(_proteinResidues)].copy()
-        S_atoms = atoms[(atoms['chainID'] == self.chain_DNA) & (atoms['name'] == 'S') & atoms['resname'].isin(_dnaResidues)].copy()
-        CA_index = [int(atom.index) for atom in CA_atoms.itertuples()]
-        if self.protein_seg: self.force.addGroup([CA_index[x] for x in self.group])
-        else:   self.force.addGroup(CA_index)
-        self.force.addGroup([int(atom.index) for atom in S_atoms.itertuples()])
-        bondGroups = [0, 1]
-        print(self.force.getGroupParameters(0))
-        print(self.force.getGroupParameters(1))
-
-        self.force.addBond(bondGroups)
-
-
-class String_length_ProteinDNA(ProteinDNAForce):
-    """ Protein-DNA string potential (Xinyu)"""
-    def __init__(self, dna, protein, chain_protein='A', chain_DNA='B', protein_seg=False, group=[], force_group=17):
-        self.force_group = force_group
-        self.chain_protein = chain_protein
-        self.chain_DNA = chain_DNA
-        self.protein_seg = protein_seg
-        self.group = group
-        super().__init__(dna, protein)
-
-    def reset(self):
-        length = openmm.CustomCentroidBondForce(2, "distance(g1,g2)")
-        length.setForceGroup(self.force_group)
-        self.force = length
-
-    def defineInteraction(self):
-        atoms = self.dna.atoms.copy()
-        atoms['index'] = atoms.index
-        CA_atoms = atoms[(atoms['chainID'] == self.chain_protein) & (atoms['name'] == 'CA') & atoms['resname'].isin(_proteinResidues)].copy()
-        S_atoms = atoms[(atoms['chainID'] == self.chain_DNA) & (atoms['name'] == 'S') & atoms['resname'].isin(_dnaResidues)].copy()
-        CA_index = [int(atom.index) for atom in CA_atoms.itertuples()]
-        if self.protein_seg: self.force.addGroup([CA_index[x] for x in self.group])
-        else:   self.force.addGroup(CA_index)
-        self.force.addGroup([int(atom.index) for atom in S_atoms.itertuples()])
-        bondGroups = [0, 1]
-        print(self.force.getGroupParameters(0))
-        print(self.force.getGroupParameters(1))
-
-        self.force.addBond(bondGroups)
-
 # List forces
 forces = dict(Bond=Bond,
               Angle=Angle,
@@ -1756,57 +701,7 @@ protein_dna_forces=dict(ExclusionProteinDNA=ExclusionProteinDNA,
                         ElectrostaticsProteinDNA=ElectrostaticsProteinDNA)
 
 
-def parse_xyz(filename=''):
-    columns = ['N', 'timestep', 'id', 'name', 'x', 'y', 'z']
-    data = []
-    with open(filename, 'r') as traj_file:
-        atom = pandas.Series(index=columns)
-        atom['id'] = None
-        for line in traj_file:
-            s = line.split()
-            if len(s) == 1:
-                atom['N'] = int(s[0])
-                if atom['id'] > -1:
-                    assert atom['id'] == atoms
-                atoms = int(s[0])
-            elif len(s) == 3:
-                atom['timestep'] = int(s[2])
-                atom['id'] = 0
-            elif len(s) > 3:
-                atom['name'] = int(s[0])
-                atom['x'], atom['y'], atom['z'] = [float(a) for a in s[1:4]]
-                data += [atom.copy()]
-                atom['id'] += 1
-    xyz_data = pandas.concat(data, axis=1).T
-    for i in ['N', 'timestep', 'id', 'name']:
-        xyz_data[i] = xyz_data[i].astype(int)
-    return xyz_data
 
-
-def parse_log(filename=''):
-    columns = ''
-    log_data = []
-    with open(filename, 'r') as log_file:
-        start = False
-        for line in log_file:
-            if line[:4] == 'Step':
-                columns = line.split()
-                start = True
-                continue
-            if start:
-                try:
-                    log_data += [[float(a) for a in line.split()]]
-                except ValueError:
-                    break
-    log_data = pandas.DataFrame(log_data, columns=columns)
-
-    try:
-        for i in ['Step', 'nbp']:
-            log_data[i] = log_data[i].astype(int)
-    except KeyError:
-        for i in ['Step', 'v_nbp']:
-            log_data[i] = log_data[i].astype(int)
-    return log_data
 
 
 # Some typical errors
